seperating summary results from version capture results into two seperate py scripts and tasks to make for a more module code
