+++ workflows/influenza_assembly_summary.wdl
@@ -18,7 +18,8 @@ workflow influenza_assembly_summary{
         Array[File] version_capture_file
 
         # python scripts
-        File summary_py
+        File results_summary_py
+        File capture_version_summary_py
     }
 
     String project_name = select_first(project_name_array)
@@ -30,7 +31,7 @@ workflow influenza_assembly_summary{
     }
 
 
-    call summary.summary as summary {
+    call summary.results_summary as results_summary {
         input:
             sample_name = sample_name,
             preprocess_qc_metrics = preprocess_qc_metrics,
@@ -38,28 +39,36 @@ workflow influenza_assembly_summary{
             assembly_qc_metrics = assembly_qc_metrics,
             na_nextclade_tsv = na_nextclade_tsv,
             ha_nextclade_tsv = ha_nextclade_tsv,
-            version_capture_file = version_capture_file,
-            workflow_version = capture_workflow_version.workflow_version,
             analysis_date = capture_workflow_version.analysis_date,
-            python_script = summary_py,
+            workflow_version = capture_workflow_version.workflow_version,
+            python_script = results_summary_py,
             project_name = project_name
         
     }
 
+        call summary.version_capture_summary as version_capture_summary {
+        input:
+            version_capture_file = version_capture_file,
+            workflow_version = capture_workflow_version.workflow_version,
+            analysis_date = capture_workflow_version.analysis_date,
+            python_script = capture_version_summary_py,
+            project_name = project_name
+        
+    }
 
     call transfer.transfer_assembly_summary_wdl as summary_transfer {
         input:
-            sequencing_results_csv = summary.sequencing_results_csv,
-            version_capture_influenza_assembly_csv = summary.version_capture_influenza_assembly_csv,
-            version_capture_influenza_assembly_summary_csv = summary.version_capture_influenza_assembly_summary_csv,
+            sequencing_results_csv = results_summary.sequencing_results_csv,
+            version_capture_influenza_assembly_csv = version_capture_summary.version_capture_influenza_assembly_csv,
+            version_capture_influenza_assembly_summary_csv = version_capture_summary.version_capture_influenza_assembly_summary_csv,
             bucket_path = out_bucket_path
     }
 
     output {
 
-        File sequencing_results_csv = summary.sequencing_results_csv
-        File version_capture_influenza_assembly_csv = summary.version_capture_influenza_assembly_csv
-        File version_capture_influenza_assembly_summary_csv = summary.version_capture_influenza_assembly_summary_csv
+        File sequencing_results_csv = results_summary.sequencing_results_csv
+        File version_capture_influenza_assembly_csv = version_capture_summary.version_capture_influenza_assembly_csv
+        File version_capture_influenza_assembly_summary_csv = version_capture_summary.version_capture_influenza_assembly_summary_csv
         String transfer_date = summary_transfer.transfer_date
 
     }
